Move document to own file
