+++ langserver/cache/cache.go
@@ -14,7 +14,6 @@
 package cache
 
 import (
-	"bytes"
 	"context"
 	"go/token"
 	"sync"
@@ -22,7 +21,6 @@ import (
 	"github.com/slrtbtfs/prometheus/promql"
 	"github.com/slrtbtfs/promql-lsp/vendored/go-tools/jsonrpc2"
 	"github.com/slrtbtfs/promql-lsp/vendored/go-tools/lsp/protocol"
-	"github.com/slrtbtfs/promql-lsp/vendored/go-tools/span"
 )
 
 // We need this so we can reserve a certain position range in the FileSet
@@ -37,20 +35,6 @@ type DocumentCache struct {
 
 	Documents   map[protocol.DocumentURI]*Document
 	DocumentsMu sync.RWMutex
-}
-
-type Document struct {
-	PosData *token.File
-	Doc     *protocol.TextDocumentItem
-	Mu      sync.RWMutex
-
-	versionCtx      context.Context
-	obsoleteVersion context.CancelFunc
-
-	CompileResult CompileResult
-
-	// Wait for this before accessing  compileResults
-	Compilers sync.WaitGroup
 }
 
 type CompileResult struct {
@@ -116,76 +100,4 @@ func (c *DocumentCache) RemoveDocument(uri protocol.DocumentURI) error {
 	c.DocumentsMu.Unlock()
 
 	return nil
-}
-
-// Set the content after an update send by the client. Must increase the version number
-func (d *Document) SetContent(content string, version float64) error {
-	d.Mu.Lock()
-	defer d.Mu.Unlock()
-
-	if version <= d.Doc.Version {
-		return jsonrpc2.NewErrorf(jsonrpc2.CodeInvalidParams, "Update to file didn't increase version number")
-	}
-
-	if len(content) > maxDocumentSize {
-		return jsonrpc2.NewErrorf(jsonrpc2.CodeInternalError, "cache/addDocument: Provided Document to large.")
-	}
-
-	d.obsoleteVersion()
-
-	d.versionCtx, d.obsoleteVersion = context.WithCancel(context.Background())
-
-	d.Doc.Text = content
-	d.Doc.Version = version
-	d.PosData.SetLinesForContent([]byte(content))
-
-	return nil
-}
-
-func (d *Document) ApplyIncrementalChanges(changes []protocol.TextDocumentContentChangeEvent, version float64) (string, error) { //nolint:lll
-	d.Mu.RLock()
-
-	if version <= d.Doc.Version {
-		return "", jsonrpc2.NewErrorf(jsonrpc2.CodeInvalidParams, "Update to file didn't increase version number")
-	}
-
-	content := []byte(d.Doc.Text)
-	uri := d.Doc.URI
-
-	d.Mu.RUnlock()
-
-	for _, change := range changes {
-		// Update column mapper along with the content.
-		converter := span.NewContentConverter(uri, content)
-		m := &protocol.ColumnMapper{
-			URI:       span.URI(d.Doc.URI),
-			Converter: converter,
-			Content:   content,
-		}
-
-		spn, err := m.RangeSpan(*change.Range)
-
-		if err != nil {
-			return "", err
-		}
-
-		if !spn.HasOffset() {
-			return "", jsonrpc2.NewErrorf(jsonrpc2.CodeInternalError, "invalid range for content change")
-		}
-
-		start, end := spn.Start().Offset(), spn.End().Offset()
-		if end < start {
-			return "", jsonrpc2.NewErrorf(jsonrpc2.CodeInternalError, "invalid range for content change")
-		}
-
-		var buf bytes.Buffer
-
-		buf.Write(content[:start])
-		buf.WriteString(change.Text)
-		buf.Write(content[end:])
-
-		content = buf.Bytes()
-	}
-
-	return string(content), nil
 }
